@@ -23,7 +23,7 @@ keywords = TensorFlow, model, fitting, scalable, HEP, zfit
 
 [options]
 packages = find:
-python_requires = >=3.8
+python_requires = >=3.9
 include_package_data = True
 testpaths = tests
 zip_safe = False
